apps/tuya.ai/your_pop_robot/tools/requirements-imu-ui.txt
pyserial>=3.5
matplotlib>=3.0
tornado>=6.0
